4A-4D exercises finished.

+++ Osa_4/blogilista/controllers/blogs.js
@@ -1,8 +1,11 @@
+const config = require('../utils/config')
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
   
@@ -15,14 +18,24 @@ blogsRouter.post('/', async (request, response, next) => {
   if(!body.title || !body.url) {
     response.status(400).end()
   } else {
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: bodyLikes
-    })
     try{
+      const decodedToken = jwt.verify(request.token, config.SECRET)    
+      if (!request.token || !decodedToken.id) {      
+        return response.status(401).json({ error: 'token missing or invalid' })    
+      }
+      const user = await User.findById(decodedToken.id)
+
+      const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: bodyLikes,
+        user: user._id
+      })
+      
       const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
       response.status(201).json()
     } catch(error) {
       next(error)
@@ -64,8 +77,18 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try{
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, config.SECRET)    
+    if (!request.token || !decodedToken.id) {      
+      return response.status(401).json({ error: 'token missing or invalid' })    
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(400).send({ error: 'blog cannot be removed by anyone else than its creator' })
+    }
   } catch(error) {
     next(error)
   }
